feat: added the Delete Account feature and Account Management feature to the porfolio settings section
